+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - New option: xtc start --port|-p [number]
 - Commands that require local xtc install exit if xtc is not found
 - Refactored all async execution to use promises.
+- `xtc help command` now works.